+++ src/parameters/parameter_audiolevel.hpp
@@ -15,14 +15,13 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ******************************************************************************/
 
-
-static bool effect_parameter_source_add(void *data, obs_source_t *source) {
+static bool effect_parameter_audiolevel_add(void *data, obs_source_t *source) {
     std::list<std::string> *sources_list = (std::list<std::string>*)(data);
 
     uint32_t flags = obs_source_get_output_flags(source);
     obs_source_type type = obs_source_get_type(source);
 
-    if ((flags & OBS_SOURCE_VIDEO) && ((type == OBS_SOURCE_TYPE_INPUT) || (type == OBS_SOURCE_TYPE_SCENE))) {
+    if ((flags & OBS_SOURCE_AUDIO) && ((type == OBS_SOURCE_TYPE_INPUT))) {
         const char *name = obs_source_get_name(source);
         if (name != NULL) {
             sources_list->push_back(std::string(name));
@@ -31,28 +30,29 @@ static bool effect_parameter_source_add(void *data, obs_source_t *source) {
     return true;
 }
 
-class effect_parameter_source : public effect_parameter {
+class effect_parameter_audiolevel : public effect_parameter {
     private:
-        gs_texrender_t *source_texrender = NULL;
         obs_source_t *source = NULL;
-
+        obs_volmeter_t *obs_volmeter;
+        float smoothing = 0.5;
 
     public:
-        effect_parameter_source(gs_eparam_t *shader_param) : effect_parameter(sizeof(float), shader_param) {
-            this->source_texrender = gs_texrender_create(GS_RGBA16, GS_ZS_NONE);
+        effect_parameter_audiolevel(gs_eparam_t *shader_param):
+            effect_parameter(sizeof(float), shader_param),
+            obs_volmeter(obs_volmeter_create(OBS_FADER_LOG))
+        {
+            obs_volmeter_add_callback(this->obs_volmeter, effect_parameter_audiolevel::callback, this);
+            obs_volmeter_set_peak_meter_type(this->obs_volmeter, SAMPLE_PEAK_METER);
+            *((float*)this->data) = 40.0;
         }
 
-        virtual ~effect_parameter_source() {
+        virtual ~effect_parameter_audiolevel() {
             if (this->source != NULL) {
                 obs_source_release(this->source);
                 this->source = NULL;
             }
-            if (this->source_texrender != NULL) {
-                obs_enter_graphics();
-                gs_texrender_destroy(this->source_texrender);
-                this->source_texrender = NULL;
-                obs_leave_graphics();
-            }
+            obs_volmeter_remove_callback(this->obs_volmeter, effect_parameter_audiolevel::callback, this);
+            obs_volmeter_destroy(this->obs_volmeter);
         }
 
         virtual void set_defaults(obs_data_t *metadata) {
@@ -60,65 +60,70 @@ class effect_parameter_source : public effect_parameter {
         }
 
         virtual void set_default(obs_data_t *settings, const char *full_param_name) {
-            UNUSED_PARAMETER(settings);
-            UNUSED_PARAMETER(full_param_name);
+            obs_data_set_default_double(settings, full_param_name, 0.5);
         }
 
         virtual void render_property_ui(const char *full_param_name, obs_properties_t *props) {
             obs_property_t *p = obs_properties_add_list(props, full_param_name, label.c_str(), OBS_COMBO_TYPE_LIST, OBS_COMBO_FORMAT_STRING);
             std::list<std::string> sources_list;
-            obs_enum_sources(effect_parameter_source_add, &sources_list);
-            obs_enum_scenes(effect_parameter_source_add, &sources_list);
+            obs_enum_sources(effect_parameter_audiolevel_add, &sources_list);
             sources_list.sort(compare_nocase);
             for (const std::string &str: sources_list) {
                 obs_property_list_add_string(p, str.c_str(), str.c_str());
             }
+            obs_properties_add_float_slider(props, (std::string(full_param_name) + "__smoothing").c_str(), "∟ Smoothing", 0.0, 0.99, 0.01);
         }
 
         virtual void set_data_from_default() {
-            // TODO check that we should maybe decrement the showing counter here ?
+            // TODO check that we should not decrement the showing counter here ?
             this->source = NULL;
         }
 
         virtual void set_data_from_settings(obs_data_t *settings, const char *full_param_name) {
             if (this->source != NULL) {
                 this->hide();
+
+                obs_volmeter_detach_source(obs_volmeter);
+
+                // TODO release earlier ?
                 obs_source_release(this->source);
                 this->source = NULL;
             }
 
             this->source = obs_get_source_by_name(obs_data_get_string(settings, full_param_name));
+            obs_volmeter_attach_source(obs_volmeter, this->source);
             if (this->source != NULL) {
                 this->show();
             }
+
+            this->smoothing = (float)obs_data_get_double(settings, (std::string(full_param_name) + "__smoothing").c_str());
         }
 
-        virtual void try_gs_set_val() {
-            if (this->source != NULL) {
-                gs_texrender_reset(this->source_texrender);
-                int cx = obs_source_get_width(this->source);
-                int cy = obs_source_get_height(this->source);
-                if (gs_texrender_begin(this->source_texrender, cx, cy)) {
-                    gs_ortho(0.0f, (float)cx, 0.0f, (float)cy, -100.0f, 100.0f); // This line took me A WHOLE WEEK to figure out
-                    obs_source_video_render(this->source);
-                    gs_texrender_end(this->source_texrender);
-                    gs_texture_t *texture = gs_texrender_get_texture(this->source_texrender);
-                    try_gs_effect_set_texture(shader_param, texture);
-                }
-            }
-        }
+        static void callback(void *data,
+          const float magnitude[MAX_AUDIO_CHANNELS],
+          const float peak[MAX_AUDIO_CHANNELS],
+          const float inputPeak[MAX_AUDIO_CHANNELS]) {
+            UNUSED_PARAMETER(magnitude);
+            UNUSED_PARAMETER(peak);
+            UNUSED_PARAMETER(inputPeak);
+            effect_parameter_audiolevel *param = static_cast<effect_parameter_audiolevel *>(data);
 
-        virtual void show() {
-            if (this->source != NULL) {
-                debug("Inc showing %s", obs_source_get_name(this->source));
-                obs_source_inc_showing(this->source);
-            }
-        }
+            float prev_val = *((float*)param->data);
 
-        virtual void hide() {
-            if (this->source != NULL) {
-                debug("Dec showing %s", obs_source_get_name(this->source));
-                obs_source_dec_showing(this->source);
+            float clamped_peak = magnitude[0];
+
+            if (clamped_peak < -100.0) {
+                clamped_peak = -100.0;
             }
+            else if (clamped_peak > 0.0) {
+                clamped_peak = 0.0;
+            }
+
+            float val = prev_val < clamped_peak ? clamped_peak : prev_val * param->smoothing + clamped_peak * (1.0 - param->smoothing);
+
+            *((float*)param->data) = (float) val;
+            debug("AUDIO LEVEL CALLBACK %f", *((float*)param->data));
         }
 };
+
+